pyproject.toml
[project]
name = "python-project-52"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "dj-database-url>=3.0.1",
    "django>=5.2.5",
    "django-bootstrap5>=25.2",
    "django-crispy-forms>=2.4",
    "gunicorn>=23.0.0",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.1",
]
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task_manager.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = ["--tb=short", "--strict-markers"]